requirements.txt
appdirs
astor
attrdict
Babel
bce-python-sdk
beautifulsoup4
black
cachetools
certifi
charset-normalizer
Cython
click
contourpy
cssselect
cssutils
cycler
decorator
dill
et-xmlfile
fire
Flask
Flask-Babel
fonttools
future
idna
imageio
imgaug
isort
itsdangerous
Jinja2
kiwisolver
lmdb
lxml
MarkupSafe
matplotlib
multiprocess
mypy-extensions
networkx
numpy
opencv-contrib-python
opencv-python
openpyxl
opt-einsum
packaging
paddle-bfloat
paddleocr
pandas
pathspec
pdf2docx
Pillow
premailer
protobuf
pyclipper
pycryptodome
PyMuPDF
pyparsing
python-dateutil
python-docx
pytz
PyWavelets
PyYAML
rapidfuzz
regex
requests
scikit-image
scipy
shapely
six
soupsieve
termcolor
tifffile
tomli
torch
torchvision
tqdm
typing_extensions
urllib3
visualdl
Werkzeug
easy-paddle-ocr
gitpython
ipython
psutil
thop
tensorboard
seaborn
albumentations
pycocotools
gradio_imageslider
gradio